wip: splitting all utils function into own files while adjusting their APIs
