+++ src/utils/escapeRegex.js
@@ -1,0 +1,3 @@
+const escapeRegex = str => str.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&')
+
+export default escapeRegex
