+++ src/utils/findPositionOfCapturingGroup.spec.js
@@ -1,3 +1,4 @@
+import expect from 'expect'
 import findPositionOfCapturingGroup from './findPositionOfCapturingGroup'
 
 describe('#findPositionOfCapturingGroup', () => {
