+++ src/utils/spliceString.js
@@ -1,0 +1,4 @@
+const spliceString = (str, start, end, insert) =>
+  str.substring(0, start) + insert + str.substring(end)
+
+export default spliceString
